# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.EMRIntegrationCreateDTO:
    properties:
      api_endpoint:
        type: string
      api_key:
        type: string
      api_secret:
        type: string
      auth_config:
        type: string
      auth_type:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      emr_system_version:
        type: string
      facility_id:
        type: integer
      notes:
        type: string
      sync_enabled:
        type: boolean
      sync_frequency:
        type: string
      webhook_url:
        type: string
    required:
    - emr_system_code
    - emr_system_name
    - facility_id
    type: object
  dto.EMRIntegrationResponseDTO:
    properties:
      api_endpoint:
        type: string
      auth_type:
        type: string
      created_at:
        type: string
      emr_system_code:
        type: string
      emr_system_name:
        type: string
      emr_system_version:
        type: string
      facility_code:
        type: string
      facility_id:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      is_verified:
        type: boolean
      last_sync_at:
        type: string
      last_sync_message:
        type: string
      last_sync_status:
        type: string
      notes:
        type: string
      sync_enabled:
        type: boolean
      sync_frequency:
        type: string
      updated_at:
        type: string
      verified_at:
        type: string
      verified_by:
        type: string
    type: object
  dto.EMRIntegrationUpdateDTO:
    properties:
      api_endpoint:
        type: string
      api_key:
        type: string
      api_secret:
        type: string
      auth_config:
        type: string
      auth_type:
        type: string
      emr_system_version:
        type: string
      is_active:
        type: boolean
      notes:
        type: string
      sync_enabled:
        type: boolean
      sync_frequency:
        type: string
      webhook_url:
        type: string
    type: object
  dto.EMRSyncLogResponseDTO:
    properties:
      completed_at:
        type: string
      duration_seconds:
        type: integer
      error_message:
        type: string
      id:
        type: integer
      integration_id:
        type: integer
      records_failed:
        type: integer
      records_processed:
        type: integer
      records_successful:
        type: integer
      started_at:
        type: string
      status:
        type: string
      sync_direction:
        type: string
      sync_type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: ict@moh.go.ug
    name: DHI
  description: Backend API for Supply Chain Repository / Bridge
  termsOfService: https://health.go.ug
  title: MoH Supply Chian Repository API
  version: "1.0"
paths:
...
  /emr-integrations:
    get:
      parameters:
      - description: Filter by facility ID
        in: query
        name: facility_id
        type: integer
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EMRIntegrationResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List EMR integrations
      tags:
      - EMRIntegration
    post:
      consumes:
      - application/json
      parameters:
      - description: Integrations payload
        in: body
        name: payload
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.EMRIntegrationCreateDTO'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.EMRIntegrationResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create EMR integrations
      tags:
      - EMRIntegration
  /emr-integrations/{id}:
    get:
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EMRIntegrationResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get EMR integration by ID
      tags:
      - EMRIntegration
    put:
      consumes:
      - application/json
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.EMRIntegrationUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EMRIntegrationResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update EMR integration
      tags:
      - EMRIntegration
  /emr-integrations/{id}/sync-logs:
    get:
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.EMRSyncLogResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get EMR sync logs
      tags:
      - EMRIntegration
  /emr-integrations/{id}/verify:
    post:
      parameters:
      - description: Integration ID
        in: path
        name: id
        required: true
        type: integer
      - description: Verify payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EMRIntegrationResponseDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify EMR integration
      tags:
      - EMRIntegration
...
swagger: "2.0"